tests/wpt/tests/html/dom/elements/global-attributes/the-translate-attribute-010.html

<!DOCTYPE html>
<html  lang="en" >
<head>
<meta charset="utf-8"/>
<title>translate inherits no</title>
<link rel='author' title='Richard Ishida' href='mailto:ishida@w3.org'>
<link rel='help' href='https://html.spec.whatwg.org/multipage/#the-translate-attribute'>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style type='text/css'>
</style>
</head>
<body>



<div class="test"><div id="box" translate="no">&nbsp; <span id="spantest">&nbsp;</span></div></div>


<script>
test(function() {
assert_false(document.getElementById('spantest').translate);
}, "If the translate attribute is set to no, javascript will detect the translation mode of elements inside that element with no translate flag as no-translate.");
</script>

<div id='log'></div>

</body>
</html>
